+++ packages/itwin/grouping-mapping-widget/src/widget/components/MappingUIActionGroup.tsx
@@ -72,10 +72,6 @@ export const MappingUIActionGroup = ({
     >
       <IconButton styleType="borderless" title='Mapping Options'>
         <SvgMore
-          style={{
-            width: "16px",
-            height: "16px",
-          }}
         />
       </IconButton>
     </DropdownMenu>
